refactor: 우리 나라에 대한 로또 패턴으로 수정
