+++ src/main/kotlin/domain/service/LottoGenerator.kt
@@ -3,22 +3,19 @@ package domain.service
 import domain.model.Lotto
 import domain.model.PurchaseLotto
 import domain.model.PurchasePrice
+import domain.strategy.KoreanLottoGenerator
+import domain.strategy.LottoCountry
 
 class LottoGenerator(
     private val money: PurchasePrice,
-) {
+    ) {
     fun makeLotto(): PurchaseLotto {
-        val purchaseLottoCount = money.value / PurchasePrice.Companion.STANDARD_AMOUNT_UNIT
-        return PurchaseLotto(List<Lotto>(purchaseLottoCount) { makeOneLotto() })
+        val purchaseLottoCount = money.value / PurchasePrice.STANDARD_AMOUNT_UNIT
+        return PurchaseLotto(List(purchaseLottoCount) { makeOneLotto(KoreanLottoGenerator()) })
     }
 
-    private fun makeOneLotto(): Lotto {
-        val newLotto = (LOTTO_MIN..LOTTO_MAX).shuffled().take(6)
+    private fun makeOneLotto(lottoCountry: LottoCountry): Lotto {
+        val newLotto = lottoCountry.generateNumber()
         return Lotto(newLotto)
-    }
-
-    companion object {
-        const val LOTTO_MIN = 1
-        const val LOTTO_MAX = 45
     }
 }
